+++ src/Coffer/Coffer.ASPNET/Controllers/ItemsController.cs
@@ -21,7 +21,7 @@ namespace Coffer.ASPNET.Controllers
             Path.Combine(Env.GetString("IMAGESTORE_PATH") ?? throw new InvalidOperationException("IMAGESTORE_PATH envionmental variable is not set"), "items");
         private readonly string tempFolder = Env.GetString("IMAGECHECK_TEMP_PATH") ?? throw new InvalidOperationException("IMAGECHECK_TEMP_PATH envionmental variable is not set");
         private readonly HttpClient _httpClient;
-        public ItemsController(IItemsRepository repository, IUsersRepository usersRepository, IImageService imageService, HttpClient httpClient) : base(repository)
+        public ItemsController(IItemsRepository repository, IUsersRepository usersRepository, IImageService imageService, IPermissionService<Guid, ItemRequired> permissionService, HttpClient httpClient) : base(repository, permissionService)
         {
             _itemsRepository = repository;
             _imageService = imageService;
@@ -48,18 +48,38 @@ namespace Coffer.ASPNET.Controllers
         [HttpPut("{id}")]
         public override async Task<ActionResult<ItemProvided>> Update(Guid id, [FromBody] ItemRequired required)
         {
+
+            if (!UserId.HasValue)
+            {
+                return Unauthorized();
+            }
+
             if (required.Quantity == 0)
             {
                 var item = await _repository.GetItemByIdAsync(id);
                 if (item == null) return NotFound();
+
+                bool isPermissionGranted = await _permissionService.CanDelete(UserId.Value, id);
+                if (!isPermissionGranted)
+                {
+                    return Forbid();
+                }
 
                 if (!string.IsNullOrEmpty(item.Image))
                     await _imageService.DeleteImageAsync(item.Image, imageFolder);
 
                 return await base.Delete(id);
             }
+            else
+            {
+                bool isPermissionGranted = await _permissionService.CanUpdate(UserId.Value, id, null);
+                if (!isPermissionGranted)
+                {
+                    return Forbid();
+                }
 
-            return await base.Update(id, required);
+                return await base.Update(id, required);
+            }
         }
 
         public record ItemIds(Guid TempId, Guid Id);
@@ -70,6 +90,12 @@ namespace Coffer.ASPNET.Controllers
         [HttpPost("Upsert")]
         public async Task<ActionResult> Upsert([FromBody] UpsertItemRequest[] request)
         {
+
+            if (!UserId.HasValue)
+            {
+                return Unauthorized();
+            }
+
             if (request == null || request.Length == 0)
                 return BadRequest("No items provided.");
 
@@ -83,14 +109,28 @@ namespace Coffer.ASPNET.Controllers
 
                 foreach (var upsertItem in request)
                 {
+
                     var itemFound = await _repository.GetItemByIdAsync(upsertItem.Id);
                     if (itemFound == null)
                     {
+
+                        bool isPermissionGranted = await _permissionService.CanCreate(UserId.Value, upsertItem.Item);
+                        if (!isPermissionGranted)
+                        {
+                            return Forbid();
+                        }
+
                         var newItem = await _repository.InsertItemAsync(upsertItem.Item);
                         newItems.Add(new InsertedItem(new ItemIds(upsertItem.Id, newItem.Id), newItem));
                     }
                     else
                     {
+                        bool isPermissionGranted = await _permissionService.CanUpdate(UserId.Value, upsertItem.Id, upsertItem.Item);
+                        if (!isPermissionGranted)
+                        {
+                            return Forbid();
+                        }
+
                         await _repository.UpdateItemAsync(upsertItem.Id, upsertItem.Item);
                     }
                 }
@@ -161,6 +201,17 @@ namespace Coffer.ASPNET.Controllers
         [HttpDelete("{id}")]
         public override async Task<ActionResult> Delete(Guid id)
         {
+
+            if (!UserId.HasValue)
+            {
+                return Unauthorized();
+            }
+            bool isPermissionGranted = await _permissionService.CanDelete(UserId.Value, id);
+            if (!isPermissionGranted)
+            {
+                return Forbid();
+            }
+
             var item = await _repository.GetItemByIdAsync(id);
             if (item == null)
                 return NotFound();
